# swagger/swagger.yaml
swagger: "2.0"
info:
  version: '1.0.0'
  title: Swagger sam-api
...
  contact:
    email: norbert.bondarczuk@wipro.com
host: localhost:8000
basePath: /api
securityDefinitions: {}
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /account:
    post:
      description: "A new mapping is created in W like Working status. Attempts to create item in any other status will be errored out. \n\nRequires:\n- Booker role."
      summary: AccountCreateOne
      tags:
      - account
      operationId: AccountCreateOne
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: body
        in: body
        required: true
        description: Creation of object Account
        schema:
          $ref: '#/definitions/RequestSetAccount'
      responses:
        201:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetAccount'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        403:
          description: Forbidden
          schema:
            $ref: '#/definitions/ResultSetError'			
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'
    get:
      description: A set of accouts is read and returned. The recoords are active ones, ie. status W or C for release 0 or status P for max release_id
      summary: AccountReadActiveAll
      tags:
      - account
      operationId: AccountReadActiveAll
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: Content-Type
        in: header
        required: true
        type: string
        enum:
        - application/json
        - application/xmlx
        - application/csv
        description: 'Determines response payload format'		
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetAccounts'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'      
    delete:
      description: Purges whole resource
      summary: AccountDeleteAll
      tags:
      - account
      operationId: AccountDeleteAll       
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetAccounts'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'                  
...
  /account/{status}/{release}/{bscsAccount}:
    put:
      description: "Most selective access to single row. The primary key is used. One record is returnd. The status and release must be given a priori. Promotion from status value W like Working to C like Controlled may be done only with Controll role.\n\nRequires:\n  \n- Booker role while in W like Work,\n  \n- Control role while in C like Control state and in transition from W like Work to C like Control."
      summary: AccountUpdateOne
      tags:
      - account
      operationId: AccountBscsAccountByStatusAndReleasePut
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: status
        in: path
        required: true
        enum:
        - W
        - C
        - P
        type: string
        description: status of the package, W for workm P for production
      - name: release
        in: path
        required: true
        enum:
        - 0
        - last
        type: string
        description: release sequential number, 0 for work, else production or last for latst version
      - name: bscsAccount
        in: path
        required: true
        type: string
        description: BSCS account code
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetAccounts'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        403:
          description: Forbidden
          schema:
            $ref: '#/definitions/ResultSetError'						
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ResultSetError'						
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'
    patch:
      description: "Most selective access to single row. The primary key is used. One record is returnd. The status and release must be given a priori. Promotion from status value W like Working to C like Controlled may be done only with Controll role.\n\nRequires:\n  \n- Booker role while in W like Work,\n  \n- Control role while in C like Control state and in transition from W like Work to C like Control."
      summary: AccountUpdateAttributes
      tags:
      - account
      operationId: AccountUpdateAttributesOne
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: status
        in: path
        required: true
        enum:
        - W
        - C
        - P
        type: string
        description: status of the package, W for workm P for production
      - name: release
        in: path
        required: true
        enum:
        - 0
        - last
        type: string
        description: release sequential number, 0 for work, else production or last for latst version
      - name: bscsAccount
        in: path
        required: true
        type: string
        description: BSCS account code
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetAccounts'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        403:
          description: Forbidden
          schema:
            $ref: '#/definitions/ResultSetError'						
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ResultSetError'									
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'      
    delete:
      description: "Most selective access to single row. The primary key is used. One record is returnd. The status and release must be given a priori. Only records in W like Working status may be manipulatd this way. Only records in W like Working status may be manipulatd this way. \n\nRequires:\n- Booker role."
      summary: AccountDeleteOne
      tags:
      - account
      operationId: AccountDeleteOne
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: status
        in: path
        required: true
        enum:
        - W
        - C
        - P
        type: string
        description: status of the package, W for work P for production
      - name: release
        in: path
        required: true
        enum:
        - 0
        - last
        type: string
        description: release sequential number, 0 for work, else production or last
      - name: bscsAccount
        in: path
        required: true
        type: string
        description: BSCS account number
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetCount'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        403:
          description: Forbidden
          schema:
            $ref: '#/definitions/ResultSetError'						
        404:
          description: Not found
          schema:
            $ref: '#/definitions/ResultSetError'									
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'
  /account/log:
    get:
      description: A set of accout logs is returned
      summary: AccountReadLog
      tags:
      - account
      operationId: AccountReadLog
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: X-Request-ID
        in: header
        required: false
        type: string
        format: uuid
        description: ''
      - name: account
        in: path
        required: true
        type: string
        description: BSCS account id for which log is produced  by triggersss
      responses:
        200:
          description: Successful operation
          schema:
            $ref: '#/definitions/ResultSetAccountLogs'
          headers: {}
        401:
          description: Not authenticated
          schema:
            $ref: '#/definitions/ResultSetError'
        500:
          description: Server error
          schema:
            $ref: '#/definitions/ResultSetError'      						
...
definitions:
...
  RequestSetAccount:
    title: RequestSetAccount
    type: object
    properties:
      data:
        $ref: '#/definitions/Account'
  ResultSetAccount:
    title: ResultSetAccount
    type: object
    properties:
      status:
        $ref: '#/definitions/Status'
      data:
        $ref: '#/definitions/Account'
  Account:
    title: Account
    example:
      status: W
      releaseId: 0
      bscsAccount: BSCSACCOUNT
      ofiSapAccount: OFISAPACCOUNT
      validFromDate: 2017-07-21
      vatCodeInd: VATCODEIND
      ofiSapWbsCode: OFISAPWBSCODE
      citMarkerVatFlag: 0
    type: object
    properties:
      status:
        type: string
      releaseId:
        type: integer
        format: int32
      bscsAccount:
        type: string
      ofiSapAccount:
        type: string
      validFromDate:
        type: string
        format: date
      vatCodeInd:
        type: string
      ofiSapWbsCode:
        type: string
      citMarkerVatFlag:
        type: integer
        format: int32
      entryDate:
        type: string
        format: date-time
      entryOwner:
        type: string
      updateDate:
        type: string
        format: date-time
      updateOwner:
        type: string
      releaseDate:
        type: string
        format: date-time
      releaseOwner:
        type: string
  ResultSetAccounts:
    title: ResultSetAccounts
    type: object
    properties:
      status:
        $ref: '#/definitions/Status'
      count:
        type: integer
        format: int32
      data:
        type: array
        items:
          $ref: '#/definitions/Account'
  AccountLog:
    title: AccountLog
    example:
      opcode: I
      opdate: 2017-07-21T17:32:28Z
      status: W
      releaseId: 0
      bscsAccount: BSCSACCOUNT
      ofiSapAccount: OFISAPACCOUNT
      validFromDate: 2017-07-01
      vatCodeInd: VATCODEIND
      ofiSapWbsCode: OFISAPWBSCODE
      citMarkerVatFlag: 0
    type: object
    properties:
      opcode:
        type: string
      opdate:
        type: string
        format: date-time
      bscsAccount:
        type: string
      ofiSapAccount:
        type: string
      validFromDate:
        type: string
        format: date
      vatCodeInd:
        type: string
      ofiSapWbsCode:
        type: string
      citMarkerVatFlag:
        type: integer
        format: int32
      entryDate:
        type: string
        format: date-time
      entryOwner:
        type: string
      updateDate:
        type: string
        format: date-time
      updateOwner:
        type: string
      releaseDate:
        type: string
        format: date-time
      releaseOwner:
        type: string		  
  ResultSetAccountLogs:
    title: ResultSetAccountLogs
    type: object
    properties:
      status:
        $ref: '#/definitions/Status'
      count:
        type: integer
        format: int32
      data:
        type: array
        items:
          $ref: '#/definitions/AccountLog'		  
...
  ResultSetError:
    title: ResultSetError
    type: object
    properties:
      status:
        $ref: '#/definitions/Status12'
      message:
        type: string
...
  ResultSetCount:
    title: ResultSetCount
    type: object
    properties:
      status:
        $ref: '#/definitions/Status'
      count:
        type: integer
        format: int32
...
  Status:
    title: Status
    example: Ok
    type: string
    enum:
    - Ok
  Status12:
    title: Status12
    example: Error
    type: string
    enum:
    - Error
...